here configure refresh will ask to select default app id
